rework to be compliant with new datamodel
