+++ cartography/data/indexes.cypher
@@ -451,6 +451,3 @@ CREATE INDEX IF NOT EXISTS FOR (n:KubernetesSecret) ON (n.lastupdated);
 CREATE INDEX IF NOT EXISTS FOR (n:KubernetesService) ON (n.id);
 CREATE INDEX IF NOT EXISTS FOR (n:KubernetesService) ON (n.name);
 CREATE INDEX IF NOT EXISTS FOR (n:KubernetesService) ON (n.lastupdated);
-CREATE INDEX IF NOT EXISTS FOR (n:LastpassUser) ON (n.id);
-CREATE INDEX IF NOT EXISTS FOR (n:LastpassUser) ON (n.email);
-CREATE INDEX IF NOT EXISTS FOR (n:LastpassUser) ON (n.lastupdated);
