--- cartography/data/jobs/cleanup/lastpass_import_cleanup.json
@@ -1,14 +1,0 @@
-{
-  "statements": [{
-    "query": "MATCH (n:LastpassUser) WHERE n.lastupdated <> $UPDATE_TAG WITH n LIMIT $LIMIT_SIZE DETACH DELETE (n)",
-    "iterative": true,
-    "iterationsize": 100
-  },
-  {
-    "query": "MATCH (n:LastpassTenant) WHERE n.lastupdated <> $UPDATE_TAG WITH n LIMIT $LIMIT_SIZE DETACH DELETE (n)",
-    "iterative": true,
-    "iterationsize": 100
-  }
-],
-  "name": "cleanup Lastpass"
-}
